Move public resource classes into ui.controller
